src/requirements.txt
pymongo~=3.11.4
pandas~=1.2.5
numpy~=1.21.0
Keras~=2.4.3
PyYAML~=5.4.1